+++ NEWS.md
@@ -1,5 +1,8 @@
 # SparseNUTS (development version)
 
+- Updated the condition factor to use HMC specific method in Langmore et al. (2019) [preprint](https://arxiv.org/abs/1905.09813) and return this to the user as `fit$condition.factor`.
+
+
 # SparseNUTS 1.0.0
 
 * Initial release